Create c-for-loop.cpp
took lot of time to do. coding part was easier but the test cases couldn't be passed because i was initialising the test cases like int a=8;int b=13;  I should have done cin>>a>>b; as input gets automatically done. So prev the bot was rejecting my code all the time.
Whoosh.